requirements.txt
pinecone==7.3.0
python-dotenv==1.1.1
google-cloud-storage==3.2.0
pandas==2.3.1
numpy==2.0.2
tqdm==4.67.1
sentence-transformers==2.2.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0 